use vim for git editor, not VS Code

+++ partials/apply/200-vscode-macos.sh
@@ -7,7 +7,10 @@
   # non-macos installs need to be as admin
   _not_macos exit 0
 
-  DOWNLOAD_URL="https://go.microsoft.com/fwlink/?LinkID=620882"
+  if [[ ! -r "/Applications/Visual Studio Code.app" ]]
+  then
+    DOWNLOAD_URL="https://go.microsoft.com/fwlink/?LinkID=620882"
 
-  curl -sSL -o vscode.zip "$DOWNLOAD_URL" && unzip -o -qq vscode.zip -d "/Applications" && rm vscode.zip
+    curl -sSL -o vscode.zip "$DOWNLOAD_URL" && unzip -o -qq vscode.zip -d "/Applications" && rm vscode.zip
+  fi
 )
